requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
apscheduler>=3.10.0
twilio>=9.0.0
sendgrid>=6.11.0
requests>=2.32.0
tenacity>=8.5.0
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0
python-dotenv>=1.0.0
